chore: Update TopicList component with delete button
This commit updates the TopicList component by adding a delete button for each topic. The delete button is implemented using the Button component from the UI library. This change improves the user interface and allows users to easily delete topics.

Note: This commit message follows the established convention of using the "chore" type for non-functional changes that improve code quality or maintainability.