feat: Add SweetAlert2 for successful topic addition
This commit adds the SweetAlert2 library to display a success message when a topic is successfully added. After the user clicks the "Add Topic" button, a SweetAlert2 modal is shown with the message "SweetAlert2 is working!". Upon closing the modal, the user is redirected to the home page.

Note: This commit message follows the established convention of using the "feat" type for new features or enhancements.