feat: Update API endpoints to use HTTPS
This commit updates the API endpoints in the code to use HTTPS instead of HTTP. This change ensures that all communication with the server is secure and encrypted. By using HTTPS, we can protect sensitive data and improve the overall security of the application.

Note: This commit message follows the established convention of using the "feat" type for new features or enhancements.